Initial commit
This came out of necessity. I wrote most of this for my personal site
mostly as a one-off challenge and then I wanted to use it for a second
site... So now it's a gem.

There are almost no useful features here, just the absolute minimum
required to get a semi-functional website.